refactor: consolidate build-logic into a single conventions module
Merge 8 build-logic sub-modules (basics, jvm, java-comment-preprocessor,
publishing, root-build, verification) plus build-logic-commons into one
conventions module. This eliminates redundant Kotlin compilation passes
during Gradle configuration.

- Reduces build-logic tasks from 92 to 18 for :postgresql:compileJava
- Cold build time drops from ~31s to ~15s (52% faster)
- Delete 6 plugins: toolchains (empty), java-library (wrapper),
  kotlin (unused), repositories/reproducible-builds/publish-to-tmp-maven-repo
  (inlined)
- All verification tools, publishing, and test config preserved
