--- build-logic-commons/gradle-plugin/src/main/kotlin/build-logic.kotlin-dsl-gradle-plugin.gradle.kts
@@ -1,21 +1,0 @@
-plugins {
-    id("java-library")
-    id("org.gradle.kotlin.kotlin-dsl") // this is 'kotlin-dsl' without version
-}
-
-tasks.validatePlugins {
-    failOnWarning.set(true)
-    enableStricterValidation.set(true)
-}
-
-// We need to figure out a version that is supported by the current JVM, and by the Kotlin Gradle plugin
-// So we settle on 21, 17, or 11 if the current JVM supports it
-listOf(21, 17, 11)
-    .firstOrNull { JavaVersion.toVersion(it) <= JavaVersion.current() }
-    ?.let { buildScriptJvmTarget ->
-        java {
-            toolchain {
-                languageVersion.set(JavaLanguageVersion.of(buildScriptJvmTarget))
-            }
-        }
-    }
