--- build-logic-commons/settings.gradle.kts
@@ -1,9 +1,0 @@
-dependencyResolutionManagement {
-    repositories {
-        gradlePluginPortal()
-    }
-}
-
-rootProject.name = "build-logic-commons"
-
-include("gradle-plugin")
