--- build-logic/basics/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    id("build-logic.kotlin-dsl-gradle-plugin")
-}
-
-dependencies {
-    implementation(project(":build-parameters"))
-}
