--- build-logic/basics/src/main/kotlin/build-logic.repositories.gradle.kts
@@ -1,20 +1,0 @@
-plugins {
-    id("build-logic.build-params")
-}
-
-repositories {
-    if (buildParameters.enableMavenLocal) {
-        mavenLocal()
-    }
-    mavenCentral {
-        mavenContent {
-            releasesOnly()
-        }
-    }
-    maven {
-        url = uri("https://oss.sonatype.org/content/repositories/snapshots/")
-        mavenContent {
-            snapshotsOnly()
-        }
-    }
-}
