+++ build-logic/conventions/src/main/kotlin/build-logic.errorprone.gradle.kts
@@ -5,7 +5,6 @@ import org.gradle.kotlin.dsl.dependencies
 
 plugins {
     id("java")
-    id("build-logic.repositories")
     id("build-logic.build-params")
 }
