+++ build-logic/conventions/src/main/kotlin/build-logic.java-published-library.gradle.kts
@@ -2,9 +2,16 @@ import com.github.vlsi.gradle.publishing.dsl.versionFromResolution
 
 plugins {
     id("build-logic.build-params")
-    id("build-logic.java-library")
-    id("build-logic.reproducible-builds")
+    id("build-logic.java")
+    id("java-library")
     id("build-logic.publish-to-central")
+}
+
+tasks.withType<AbstractArchiveTask>().configureEach {
+    isPreserveFileTimestamps = false
+    isReproducibleFileOrder = true
+    dirPermissions { unix("755") }
+    filePermissions { unix("644") }
 }
 
 java {
