+++ build-logic/conventions/src/main/kotlin/build-logic.without-type-annotations.gradle.kts
