--- build-logic/java-comment-preprocessor/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    id("build-logic.kotlin-dsl-gradle-plugin")
-}
-
-dependencies {
-    implementation("com.igormaznitsa:jcp:7.3.0")
-    implementation("com.github.vlsi.gradle-extensions:com.github.vlsi.gradle-extensions.gradle.plugin:3.0.2")
-}
