--- build-logic/jvm/src/main/kotlin/build-logic.java-library.gradle.kts
@@ -1,4 +1,0 @@
-plugins {
-    id("build-logic.java")
-    id("java-library")
-}
