--- build-logic/publishing/src/main/kotlin/build-logic.publish-to-tmp-maven-repo.gradle.kts
@@ -1,41 +1,0 @@
-plugins {
-    id("java-library")
-    id("maven-publish")
-}
-
-val localRepoElements by configurations.creating {
-    isCanBeConsumed = true
-    isCanBeResolved = false
-    description =
-        "Shares local maven repository directory that contains the artifacts produced by the current project"
-    attributes {
-        attribute(Category.CATEGORY_ATTRIBUTE, objects.named("maven-repository"))
-        attribute(Bundling.BUNDLING_ATTRIBUTE, objects.named(Bundling.EXTERNAL))
-    }
-}
-
-val localRepoDir = layout.buildDirectory.dir("local-maven-repo")
-
-publishing {
-    repositories {
-        maven {
-            name = "tmp-maven"
-            url = uri(localRepoDir)
-        }
-    }
-}
-
-localRepoElements.outgoing.artifact(localRepoDir) {
-    builtBy(tasks.named("publishAllPublicationsToTmp-mavenRepository"))
-}
-
-val cleanLocalRepository by tasks.registering(Delete::class) {
-    description = "Clears local-maven-repo so timestamp-based snapshot artifacts do not consume space"
-    delete(localRepoDir)
-}
-
-tasks.withType<PublishToMavenRepository>()
-    .matching { it.name.endsWith("PublicationToTmp-mavenRepository") }
-    .configureEach {
-        dependsOn(cleanLocalRepository)
-    }
