+++ build-logic/settings.gradle.kts
@@ -11,11 +11,5 @@ dependencyResolutionManagement {
 
 rootProject.name = "build-logic"
 
-includeBuild("../build-logic-commons")
 include("build-parameters")
-include("basics")
-include("jvm")
-include("java-comment-preprocessor")
-include("publishing")
-include("root-build")
-include("verification")
+include("conventions")
